+++ base/exports.jl
@@ -77,7 +77,6 @@ export
     Matrix,
     MergeSort,
     NTuple,
-    Nullable,
     ObjectIdDict,
     OrdinalRange,
     Pair,
@@ -101,6 +100,7 @@ export
     AbstractSerializer,
     SerializationState,
     Set,
+    Some,
     StepRange,
     StepRangeLen,
     StridedArray,
@@ -150,7 +150,6 @@ export
     EOFError,
     InvalidStateException,
     KeyError,
-    NullException,
     ParseError,
     SystemError,
     UnicodeError,
@@ -1158,10 +1157,6 @@ export
     unsafe_read,
     unsafe_store!,
     unsafe_write,
-
-# nullable types
-    isnull,
-    unsafe_get,
 
 # Macros
     # parser internal