+++ base/sysimg.jl
@@ -85,7 +85,6 @@ end
 include("essentials.jl")
 include("ctypes.jl")
 include("gcutils.jl")
-include("nullabletype.jl")
 include("generator.jl")
 include("reflection.jl")
 include("options.jl")
@@ -250,8 +249,8 @@ include("multidimensional.jl")
 include("permuteddimsarray.jl")
 using .PermutedDimsArrays
 
-# nullable types
-include("nullable.jl")
+# Some type
+include("some.jl")
 
 include("broadcast.jl")
 using .Broadcast